fix(cockpit): delegate Rust complexity analysis
