+++ ci/proof.toml
@@ -268,6 +268,7 @@ kind = "rust"
 paths = [
   "crates/tokmd-analysis/src/complexity/**",
   "crates/tokmd-analysis/src/content/complexity.rs",
+  "crates/tokmd-analysis/src/source_complexity.rs",
 ]
 packages = ["tokmd-analysis"]
 proof = ["cargo test -p tokmd-analysis --all-features complexity --verbose"]
