+++ crates/tokmd-cockpit/Cargo.toml
@@ -22,6 +22,7 @@ ignore = "0.4.23"
 serde.workspace = true
 serde_json.workspace = true
 time = { version = "0.3.41", features = ["formatting"] }
+tokmd-analysis = { path = "../tokmd-analysis", version = "1.10.0", default-features = false }
 tokmd-analysis-types.workspace = true
 tokmd-envelope.workspace = true
 tokmd-types.workspace = true
@@ -33,4 +34,3 @@ tokmd-git = { workspace = true, optional = true }
 insta = { workspace = true }
 proptest.workspace = true
 tempfile.workspace = true
-
